Private variables/methods made protected
Since this project is no longer maintained, people who would still like to use it should create a child class of the original implementation and make changes in the child class. To do this, we need to open up private declarations as protected.

For example, the PAYPAL_HOST & SANDBOX_HOST used in this project are incorrect (should be ipnpb.paypal.com & ipnpb.sandbox.paypal.com). Since these are const values, they cannot be changed even in child classes. However, we could override the getPaypalHost method in a child class and make it return the correct urls.